refactor: enum 통합
